Migrate web client to apps/client
