+++ apps/client/.env.local.template
