+++ apps/client/.prettierignore
