+++ apps/client/.prettierrc
