+++ apps/client/src/features/landing/components/footer.tsx
