+++ apps/client/src/features/landing/components/hero.tsx
