+++ apps/client/src/libs/ui/echo-logo.tsx
