+++ apps/client/src/providers/auth-provider.tsx
