+++ apps/client/src/routes/(protected)/_layout.tsx
