+++ apps/client/src/routes/(public)/(auth)/_layout.tsx
