+++ apps/client/src/routes/(public)/(auth)/register.tsx
