+++ apps/client/src/routes/(public)/(auth)/sso-callback.tsx
