+++ apps/client/src/routes/(public)/_layout.tsx
