+++ apps/client/src/routes/(public)/index.tsx
