+++ apps/client/src/types/global.d.ts
