+++ apps/client/src/utils/auth.ts
