--- client/src/routes/(protected)/home.tsx
@@ -1,15 +1,0 @@
-import { UserButton } from "@clerk/clerk-react"
-import { createFileRoute } from "@tanstack/react-router"
-
-export const Route = createFileRoute("/(protected)/home")({
-    component: HomePage
-})
-
-function HomePage() {
-    return (
-        <div className="flex h-full items-center justify-center gap-6">
-            Hello "/(protected)/home"!
-            <UserButton />
-        </div>
-    )
-}
